moved n moodified  useWalk hook logic to Player
